moving endpoints to respective traits
