+++ Cargo.toml
@@ -30,6 +30,7 @@ rspotify-http = { path = "rspotify-http", version = "0.10.0" }
 
 ### Client ###
 async-stream = { version = "0.3.0", optional = true }
+async-trait = { version = "0.1.48", optional = true }
 chrono = { version = "0.4.13", features = ["serde", "rustc-serialize"] }
 derive_builder = "0.10.0"
 dotenv = { version = "0.15.0", optional = true }
