+++ rspotify-http/Cargo.toml
@@ -49,72 +49,3 @@ ureq-rustls-tls = ["ureq/tls"]
 # Internal features for checking async or sync compilation
 __async = ["async-trait", "futures"]
 __sync = ["maybe-async/is_sync"]
-
-[package.metadata.docs.rs]
-# Also documenting the CLI methods
-features = ["cli"]
-
-[[example]]
-name = "album"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/album.rs"
-
-[[example]]
-name = "current_user_recently_played"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/current_user_recently_played.rs"
-
-[[example]]
-name = "oauth_tokens"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/oauth_tokens.rs"
-
-[[example]]
-name = "track"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/track.rs"
-
-[[example]]
-name = "tracks"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/tracks.rs"
-
-[[example]]
-name = "with_refresh_token"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/with_refresh_token.rs"
-
-[[example]]
-name = "device"
-required-features = ["env-file", "cli", "client-ureq"]
-path = "examples/ureq/device.rs"
-
-[[example]]
-name = "me"
-required-features = ["env-file", "cli", "client-ureq"]
-path = "examples/ureq/me.rs"
-
-[[example]]
-name = "search"
-required-features = ["env-file", "cli", "client-ureq"]
-path = "examples/ureq/search.rs"
-            
-[[example]]
-name = "seek_track"
-required-features = ["env-file", "cli", "client-ureq"]
-path = "examples/ureq/seek_track.rs"
-
-[[example]]
-name = "pagination_manual"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/pagination_manual.rs"
-
-[[example]]
-name = "pagination_sync"
-required-features = ["env-file", "cli", "client-ureq"]
-path = "examples/pagination_sync.rs"
-
-[[example]]
-name = "pagination_async"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/pagination_async.rs"
