+++ src/oauth2.rs
@@ -1,211 +1,5 @@
 //! User authorization and client credentials management.
 
-use chrono::prelude::*;
-use derive_builder::Builder;
-use getrandom::getrandom;
-use maybe_async::maybe_async;
-use serde::{Deserialize, Serialize};
-use url::Url;
-
-use chrono::Duration;
-use std::collections::{HashMap, HashSet};
-use std::{
-    env, fs,
-    io::{Read, Write},
-    path::Path,
-};
-
-use crate::client::{ClientResult, Spotify};
-use crate::http::{headers, Form, Headers};
-
-/// Generate `length` random chars
-pub(in crate) fn generate_random_string(length: usize) -> String {
-    let alphanum: &[u8] =
-        "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".as_bytes();
-    let mut buf = vec![0u8; length];
-    getrandom(&mut buf).unwrap();
-    let range = alphanum.len();
-
-    buf.iter()
-        .map(|byte| alphanum[*byte as usize % range] as char)
-        .collect()
-}
-
-mod auth_urls {
-    pub const AUTHORIZE: &str = "https://accounts.spotify.com/authorize";
-    pub const TOKEN: &str = "https://accounts.spotify.com/api/token";
-}
-
-mod duration_second {
-    use chrono::Duration;
-    use serde::{de, Deserialize, Serializer};
-
-    /// Deserialize `chrono::Duration` from seconds (represented as u64)
-    pub(in crate) fn deserialize<'de, D>(d: D) -> Result<Duration, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        let duration: i64 = Deserialize::deserialize(d)?;
-        Ok(Duration::seconds(duration))
-    }
-
-    /// Serialize `chrono::Duration` to seconds (represented as u64)
-    pub(in crate) fn serialize<S>(x: &Duration, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        s.serialize_i64(x.num_seconds())
-    }
-}
-
-mod space_separated_scope {
-    use serde::{de, Deserialize, Serializer};
-    use std::collections::HashSet;
-
-    pub(crate) fn deserialize<'de, D>(d: D) -> Result<HashSet<String>, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        let scope: &str = Deserialize::deserialize(d)?;
-        Ok(scope.split_whitespace().map(|x| x.to_owned()).collect())
-    }
-
-    pub(crate) fn serialize<S>(scope: &HashSet<String>, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let scope = scope.clone().into_iter().collect::<Vec<_>>().join(" ");
-        s.serialize_str(&scope)
-    }
-}
-
-/// Spotify access token information
-/// [Reference](https://developer.spotify.com/documentation/general/guides/authorization-guide/)
-#[derive(Builder, Clone, Debug, Serialize, Deserialize)]
-pub struct Token {
-    /// An access token that can be provided in subsequent calls
-    #[builder(setter(into))]
-    pub access_token: String,
-    /// The time period for which the access token is valid.
-    #[builder(default = "Duration::seconds(0)")]
-    #[serde(with = "duration_second")]
-    pub expires_in: Duration,
-    /// The valid time for which the access token is available represented
-    /// in ISO 8601 combined date and time.
-    #[builder(setter(strip_option), default = "Some(Utc::now())")]
-    pub expires_at: Option<DateTime<Utc>>,
-    /// A token that can be sent to the Spotify Accounts service
-    /// in place of an authorization code
-    #[builder(setter(into, strip_option), default)]
-    pub refresh_token: Option<String>,
-    /// A list of [scopes](https://developer.spotify.com/documentation/general/guides/scopes/)
-    /// which have been granted for this `access_token`
-    /// You could use macro [scopes!](crate::scopes) to build it at compile time easily
-    #[builder(default)]
-    #[serde(default, with = "space_separated_scope")]
-    pub scope: HashSet<String>,
-}
-
-impl TokenBuilder {
-    /// Tries to initialize the token from a cache file.
-    pub fn from_cache<T: AsRef<Path>>(path: T) -> Self {
-        if let Ok(mut file) = fs::File::open(path) {
-            let mut tok_str = String::new();
-            if file.read_to_string(&mut tok_str).is_ok() {
-                if let Ok(tok) = serde_json::from_str::<Token>(&tok_str) {
-                    return TokenBuilder {
-                        access_token: Some(tok.access_token),
-                        expires_in: Some(tok.expires_in),
-                        expires_at: Some(tok.expires_at),
-                        refresh_token: Some(tok.refresh_token),
-                        scope: Some(tok.scope),
-                    };
-                }
-            }
-        }
-
-        TokenBuilder::default()
-    }
-}
-
-impl Token {
-    /// Saves the token information into its cache file.
-    pub fn write_cache<T: AsRef<Path>>(&self, path: T) -> ClientResult<()> {
-        let token_info = serde_json::to_string(&self)?;
-
-        let mut file = fs::OpenOptions::new().write(true).create(true).open(path)?;
-        file.set_len(0)?;
-        file.write_all(token_info.as_bytes())?;
-
-        Ok(())
-    }
-
-    /// Check if the token is expired
-    pub fn is_expired(&self) -> bool {
-        self.expires_at
-            .map_or(true, |x| Utc::now().timestamp() > x.timestamp())
-    }
-}
-
-/// Simple client credentials object for Spotify.
-#[derive(Builder, Debug, Default, Clone, Serialize, Deserialize)]
-pub struct Credentials {
-    #[builder(setter(into))]
-    pub id: String,
-    #[builder(setter(into))]
-    pub secret: String,
-}
-
-impl CredentialsBuilder {
-    /// Parses the credentials from the environment variables
-    /// `RSPOTIFY_CLIENT_ID` and `RSPOTIFY_CLIENT_SECRET`. You can optionally
-    /// activate the `env-file` feature in order to read these variables from
-    /// a `.env` file.
-    pub fn from_env() -> Self {
-        #[cfg(feature = "env-file")]
-        {
-            dotenv::dotenv().ok();
-        }
-
-        CredentialsBuilder {
-            id: env::var("RSPOTIFY_CLIENT_ID").ok(),
-            secret: env::var("RSPOTIFY_CLIENT_SECRET").ok(),
-        }
-    }
-}
-
-/// Structure that holds the required information for requests with OAuth.
-#[derive(Builder, Debug, Default, Clone, Serialize, Deserialize)]
-pub struct OAuth {
-    #[builder(setter(into))]
-    pub redirect_uri: String,
-    /// The state is generated by default, as suggested by the OAuth2 spec:
-    /// [Cross-Site Request Forgery](https://tools.ietf.org/html/rfc6749#section-10.12)
-    #[builder(setter(into), default = "generate_random_string(16)")]
-    pub state: String,
-    /// You could use macro [scopes!](crate::scopes) to build it at compile time easily
-    #[builder(default)]
-    pub scope: HashSet<String>,
-    #[builder(setter(into, strip_option), default)]
-    pub proxies: Option<String>,
-}
-
-impl OAuthBuilder {
-    /// Parses the credentials from the environment variable
-    /// `RSPOTIFY_REDIRECT_URI`. You can optionally activate the `env-file`
-    /// feature in order to read these variables from a `.env` file.
-    pub fn from_env() -> Self {
-        #[cfg(feature = "env-file")]
-        {
-            dotenv::dotenv().ok();
-        }
-
-        OAuthBuilder {
-            redirect_uri: env::var("RSPOTIFY_REDIRECT_URI").ok(),
-            ..Default::default()
-        }
-    }
-}
 
 /// Authorization-related methods for the client.
 impl Spotify {
